Update assembly metadata and versioning details
Updated `AssemblyDescription` to provide a more descriptive explanation of the assembly. Updated `AssemblyCompany` to "Spectrum" and `AssemblyProduct` to "Spectrum Software Management V1" for better clarity. Changed `AssemblyCopyright` to "Copyright © Vision^n 2026."

Bumped `AssemblyVersion` from `1.0.0.0` to `4.7.2.1` to reflect a significant version update. Updated `AssemblyFileVersion` from `1.0.0.0` to `1.0.1.2` and later to `1.0.1.17` to track minor and patch-level changes.